code cleanup and switched to smaller datatypes

+++ digitaltechnik/V2/src/DmxCommand.h
@@ -1,8 +1,8 @@
 #ifndef DmxCommand_h
 #define DmxCommand_h
 
-struct DmxCommand{
-    unsigned long executionTime;
+struct DmxCommand {
+    unsigned int executionTime;
     int function;
     unsigned char value;
 };
